requirements.txt
ultralytics>=8.3.0
opencv-python
numpy
Pillow
